pyproject.toml
[project]
name = "pj03"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "easyocr>=1.7.2",
    "fastapi[standard]>=0.135.1",
    "google-cloud-vision>=3.12.1",
    "google-genai>=1.68.0",
    "huggingface-hub>=1.7.1",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.8",
    "mediapipe==0.10.14",
    "openai>=2.29.0",
    "opencv-contrib-python>=4.13.0.92",
    "opencv-python>=4.13.0.92",
    "pillow>=12.1.1",
    "python-dotenv>=1.2.2",
    "roboflow>=1.2.16",
    "scikit-learn>=1.8.0",
    "streamlit>=1.55.0",
    "torch>=2.10.0",
    "torchaudio>=2.10.0",
    "torchvision>=0.25.0",
    "transformers>=5.3.0",
    "translate>=3.8.0",
    "ultralytics>=8.4.22",
    "xgboost>=3.2.0",
]


[tool.uv.sources]
torch = { index = "pytorch" }
torchvision = { index = "pytorch" }

[[tool.uv.index]]
name = "pytorch"
url = "https://download.pytorch.org/whl/cu126"
explicit = true
